removing files with stuff i haven't really done much on, moving utilities to their own folder to clean things up
