+++ all_headers.hpp
@@ -3,7 +3,5 @@
 #include "src/complex_src/complex_number.hpp"
 #include "src/machine_learning_src/k_means.hpp"
 #include "src/machine_learning_src/linear_regression.hpp"
-#include "src/utility_functions.hpp"
-#include "src/recombining_tree.hpp"
-#include "src/parameterized_function.hpp"
-#include "src/ndarray.hpp"
+#include "src/utilities_src/utility_functions.hpp"
+#include "src/utilities_src/parameterized_function.hpp"
