+++ src/complex_src/complex_number.hpp
@@ -4,8 +4,8 @@
 #define complex_numbers_12903u1290
 
 #include <vector>
-#include "../../all_headers.hpp" // get rid of this
-#include "../utility_functions.hpp"
+#include "../utilities_src/utility_functions.hpp"
+#include "../doublevec_src/doublevec.hpp"
 
 class Complex
 {
