+++ src/complex_src/complex_utilities.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include <iostream>
 #include "complex_number.hpp"
-#include "../utility_functions.hpp"
+#include "../utilities_src/utility_functions.hpp"
 using namespace std;
 
 double Complex::magnitude() const
