+++ src/doublevec_src/doublevec_initializations.cpp
@@ -3,7 +3,7 @@
 #include <vector>
 #include <sstream>
 #include "doublevec.hpp"
-#include "../utility_functions.hpp"
+#include "../utilities_src/utility_functions.hpp"
 
 using namespace std;
