+++ src/matrix_src/matrix_inverse.cpp
@@ -1,5 +1,5 @@
 #include "matrix.hpp"
-#include "../utility_functions.hpp"
+#include "../utilities_src/utility_functions.hpp"
 using namespace std;
 using namespace util;
